Crate mpl_token_metadata

Modules

accounts

This code was AUTOGENERATED using the kinobi library. Please DO NOT EDIT THIS FILE, instead use visitors to add features, then rerun kinobi to update it.

errors

This code was AUTOGENERATED using the kinobi library. Please DO NOT EDIT THIS FILE, instead use visitors to add features, then rerun kinobi to update it.

hooked

instructions

This code was AUTOGENERATED using the kinobi library. Please DO NOT EDIT THIS FILE, instead use visitors to add features, then rerun kinobi to update it.

programs

This code was AUTOGENERATED using the kinobi library. Please DO NOT EDIT THIS FILE, instead use visitors to add features, then rerun kinobi to update it.

types

This code was AUTOGENERATED using the kinobi library. Please DO NOT EDIT THIS FILE, instead use visitors to add features, then rerun kinobi to update it.

utils

Constants

EDITION_MARKER_BIT_SIZE

Number of bits used by a edition marker.

ID

mpl_token_metadata program ID.

MAX_CREATOR_LEN

Maximum number of bytes used by a creator data.

MAX_CREATOR_LIMIT

Maximum number of creators in a metadata.

MAX_EDITION_MARKER_SIZE

Maximum number of bytes used by a edition marker.

MAX_NAME_LENGTH

Maximum number of characters in a metadata name.

MAX_SYMBOL_LENGTH

Maximum number of characters in a metadata symbol.

MAX_URI_LENGTH

Maximum number of characters in a metadata uri.