Enum mpl_token_metadata::error::MetadataError

pub enum MetadataError {

    InstructionUnpackError,
    InstructionPackError,
    NotRentExempt,
    AlreadyInitialized,
    Uninitialized,
    InvalidMetadataKey,
    InvalidEditionKey,
    UpdateAuthorityIncorrect,
    UpdateAuthorityIsNotSigner,
    NotMintAuthority,
    InvalidMintAuthority,
    NameTooLong,
    SymbolTooLong,
    UriTooLong,
    UpdateAuthorityMustBeEqualToMetadataAuthorityAndSigner,
    MintMismatch,
    EditionsMustHaveExactlyOneToken,
    MaxEditionsMintedAlready,
    TokenMintToFailed,
    MasterRecordMismatch,
    DestinationMintMismatch,
    EditionAlreadyMinted,
    PrintingMintDecimalsShouldBeZero,
    OneTimePrintingAuthorizationMintDecimalsShouldBeZero,
    EditionMintDecimalsShouldBeZero,
    TokenBurnFailed,
    TokenAccountOneTimeAuthMintMismatch,
    DerivedKeyInvalid,
    PrintingMintMismatch,
    OneTimePrintingAuthMintMismatch,
    TokenAccountMintMismatch,
    TokenAccountMintMismatchV2,
    NotEnoughTokens,
    PrintingMintAuthorizationAccountMismatch,
    AuthorizationTokenAccountOwnerMismatch,
    Disabled,
    CreatorsTooLong,
    CreatorsMustBeAtleastOne,
    MustBeOneOfCreators,
    NoCreatorsPresentOnMetadata,
    CreatorNotFound,
    InvalidBasisPoints,
    PrimarySaleCanOnlyBeFlippedToTrue,
    OwnerMismatch,
    NoBalanceInAccountForAuthorization,
    ShareTotalMustBe100,
    ReservationExists,
    ReservationDoesNotExist,
    ReservationNotSet,
    ReservationAlreadyMade,
    BeyondMaxAddressSize,
    NumericalOverflowError,
    ReservationBreachesMaximumSupply,
    AddressNotInReservation,
    CannotVerifyAnotherCreator,
    CannotUnverifyAnotherCreator,
    SpotMismatch,
    IncorrectOwner,
    PrintingWouldBreachMaximumSupply,
    DataIsImmutable,
    DuplicateCreatorAddress,
    ReservationSpotsRemainingShouldMatchTotalSpotsAtStart,
    InvalidTokenProgram,
    DataTypeMismatch,
    BeyondAlottedAddressSize,
    ReservationNotComplete,
    TriedToReplaceAnExistingReservation,
    InvalidOperation,
    InvalidOwner,
    PrintingMintSupplyMustBeZeroForConversion,
    OneTimeAuthMintSupplyMustBeZeroForConversion,
    InvalidEditionIndex,
    ReservationArrayShouldBeSizeOne,
    IsMutableCanOnlyBeFlippedToFalse,
    CollectionCannotBeVerifiedInThisInstruction,
    Removed,
    MustBeBurned,
    InvalidUseMethod,
    CannotChangeUseMethodAfterFirstUse,
    CannotChangeUsesAfterFirstUse,
    CollectionNotFound,
    InvalidCollectionUpdateAuthority,
    CollectionMustBeAUniqueMasterEdition,
    UseAuthorityRecordAlreadyExists,
    UseAuthorityRecordAlreadyRevoked,
    Unusable,
    NotEnoughUses,

}

Errors that may be returned by the Metadata program.

Variants

InstructionUnpackError

Failed to unpack instruction data

InstructionPackError

Failed to pack instruction data

NotRentExempt

Lamport balance below rent-exempt threshold.

AlreadyInitialized

Already initialized

Uninitialized

Uninitialized

InvalidMetadataKey

Metadata’s key must match seed of [‘metadata’, program id, mint] provided

InvalidEditionKey

Edition’s key must match seed of [‘metadata’, program id, name, ‘edition’] provided

UpdateAuthorityIncorrect

Update Authority given does not match

UpdateAuthorityIsNotSigner

Update Authority needs to be signer to update metadata

NotMintAuthority

You must be the mint authority and signer on this transaction

InvalidMintAuthority

Mint authority provided does not match the authority on the mint

NameTooLong

Name too long

SymbolTooLong

Symbol too long

UriTooLong

URI too long

UpdateAuthorityMustBeEqualToMetadataAuthorityAndSigner

Update authority must be equivalent to the metadata’s authority and also signer of this transaction

MintMismatch

Mint given does not match mint on Metadata

EditionsMustHaveExactlyOneToken

Editions must have exactly one token

MaxEditionsMintedAlready

Maximum editions printed already

TokenMintToFailed

Token mint to failed

MasterRecordMismatch

The master edition record passed must match the master record on the edition given

DestinationMintMismatch

The destination account does not have the right mint

EditionAlreadyMinted

An edition can only mint one of its kind!

PrintingMintDecimalsShouldBeZero

Printing mint decimals should be zero

OneTimePrintingAuthorizationMintDecimalsShouldBeZero

OneTimePrintingAuthorizationMint mint decimals should be zero

EditionMintDecimalsShouldBeZero

Edition mint decimals should be zero

TokenBurnFailed

Token burn failed

TokenAccountOneTimeAuthMintMismatch

The One Time authorization mint does not match that on the token account!

DerivedKeyInvalid

Derived key invalid

PrintingMintMismatch

The Printing mint does not match that on the master edition!

OneTimePrintingAuthMintMismatch

The One Time Printing Auth mint does not match that on the master edition!

TokenAccountMintMismatch

The mint of the token account does not match the Printing mint!

TokenAccountMintMismatchV2

The mint of the token account does not match the master metadata mint!

NotEnoughTokens

Not enough tokens to mint a limited edition

PrintingMintAuthorizationAccountMismatch

The mint on your authorization token holding account does not match your Printing mint!

AuthorizationTokenAccountOwnerMismatch

The authorization token account has a different owner than the update authority for the master edition!

Disabled

This feature is currently disabled.

CreatorsTooLong

Creators list too long

CreatorsMustBeAtleastOne

Creators must be at least one if set

MustBeOneOfCreators

If using a creators array, you must be one of the creators listed

NoCreatorsPresentOnMetadata

This metadata does not have creators

CreatorNotFound

This creator address was not found

InvalidBasisPoints

Basis points cannot be more than 10000

PrimarySaleCanOnlyBeFlippedToTrue

Primary sale can only be flipped to true and is immutable

OwnerMismatch

Owner does not match that on the account given

NoBalanceInAccountForAuthorization

This account has no tokens to be used for authorization

ShareTotalMustBe100

Share total must equal 100 for creator array

ReservationExists

This reservation list already exists!

ReservationDoesNotExist

This reservation list does not exist!

ReservationNotSet

This reservation list exists but was never set with reservations

ReservationAlreadyMade

This reservation list has already been set!

BeyondMaxAddressSize

Provided more addresses than max allowed in single reservation

NumericalOverflowError

NumericalOverflowError

ReservationBreachesMaximumSupply

This reservation would go beyond the maximum supply of the master edition!

AddressNotInReservation

Address not in reservation!

CannotVerifyAnotherCreator

You cannot unilaterally verify another creator, they must sign

CannotUnverifyAnotherCreator

You cannot unilaterally unverify another creator

SpotMismatch

In initial reservation setting, spots remaining should equal total spots

IncorrectOwner

Incorrect account owner

PrintingWouldBreachMaximumSupply

printing these tokens would breach the maximum supply limit of the master edition

DataIsImmutable

Data is immutable

DuplicateCreatorAddress

No duplicate creator addresses

ReservationSpotsRemainingShouldMatchTotalSpotsAtStart

Reservation spots remaining should match total spots when first being created

InvalidTokenProgram

Invalid token program

DataTypeMismatch

Data type mismatch

BeyondAlottedAddressSize

Beyond alotted address size in reservation!

ReservationNotComplete

The reservation has only been partially alotted

TriedToReplaceAnExistingReservation

You cannot splice over an existing reservation!

InvalidOperation

Invalid operation

InvalidOwner

Invalid owner

PrintingMintSupplyMustBeZeroForConversion

Printing mint supply must be zero for conversion

OneTimeAuthMintSupplyMustBeZeroForConversion

One Time Auth mint supply must be zero for conversion

InvalidEditionIndex

You tried to insert one edition too many into an edition mark pda

ReservationArrayShouldBeSizeOne

IsMutableCanOnlyBeFlippedToFalse

Is Mutable can only be flipped to false

CollectionCannotBeVerifiedInThisInstruction

Removed

MustBeBurned

InvalidUseMethod

CannotChangeUseMethodAfterFirstUse

CannotChangeUsesAfterFirstUse

CollectionNotFound

InvalidCollectionUpdateAuthority

CollectionMustBeAUniqueMasterEdition

UseAuthorityRecordAlreadyExists

UseAuthorityRecordAlreadyRevoked

Unusable

NotEnoughUses