Enum mpl_token_metadata::instruction::MetadataInstruction

pub enum MetadataInstruction {

    CreateMetadataAccount(CreateMetadataAccountArgs),
    UpdateMetadataAccount(UpdateMetadataAccountArgs),
    DeprecatedCreateMasterEdition(CreateMasterEditionArgs),
    DeprecatedMintNewEditionFromMasterEditionViaPrintingToken,
    UpdatePrimarySaleHappenedViaToken,
    DeprecatedSetReservationList(SetReservationListArgs),
    DeprecatedCreateReservationList,
    SignMetadata,
    DeprecatedMintPrintingTokensViaToken(MintPrintingTokensViaTokenArgs),
    DeprecatedMintPrintingTokens(MintPrintingTokensViaTokenArgs),
    CreateMasterEdition(CreateMasterEditionArgs),
    MintNewEditionFromMasterEditionViaToken(MintNewEditionFromMasterEditionViaTokenArgs),
    ConvertMasterEditionV1ToV2,
    MintNewEditionFromMasterEditionViaVaultProxy(MintNewEditionFromMasterEditionViaTokenArgs),
    PuffMetadata,
    UpdateMetadataAccountV2(UpdateMetadataAccountArgsV2),
    CreateMetadataAccountV2(CreateMetadataAccountArgsV2),
    CreateMasterEditionV3(CreateMasterEditionArgs),
    VerifyCollection,
    Utilize(UtilizeArgs),
    ApproveUseAuthority(ApproveUseAuthorityArgs),
    RevokeUseAuthority,
    UnverifyCollection,

}

Instructions supported by the Metadata program.

Variants

CreateMetadataAccount(CreateMetadataAccountArgs)

Create Metadata object. 0. [writable] Metadata key (pda of [‘metadata’, program id, mint id])

[] Mint of token asset

[signer] Mint authority

[signer] payer

[] update authority info

[] System program

[] Rent info

Tuple Fields of CreateMetadataAccount

0: CreateMetadataAccountArgs

UpdateMetadataAccount(UpdateMetadataAccountArgs)

Update a Metadata 0. [writable] Metadata account

[signer] Update authority key

Tuple Fields of UpdateMetadataAccount

0: UpdateMetadataAccountArgs

DeprecatedCreateMasterEdition(CreateMasterEditionArgs)

Register a Metadata as a Master Edition V1, which means Editions can be minted. Henceforth, no further tokens will be mintable from this primary mint. Will throw an error if more than one token exists, and will throw an error if less than one token exists in this primary mint. 0. [writable] Unallocated edition V1 account with address as pda of [‘metadata’, program id, mint, ‘edition’]

[writable] Metadata mint

[writable] Printing mint - A mint you control that can mint tokens that can be exchanged for limited editions of your master edition via the MintNewEditionFromMasterEditionViaToken endpoint

[writable] One time authorization printing mint - A mint you control that prints tokens that gives the bearer permission to mint any number of tokens from the printing mint one time via an endpoint with the token-metadata program for your metadata. Also burns the token.

[signer] Current Update authority key

[signer] Printing mint authority - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY.

[signer] Mint authority on the metadata’s mint - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[] Metadata account

[signer] payer

[] Token program

[] System program

[] Rent info

[signer] One time authorization printing mint authority - must be provided if using max supply. THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY.

Tuple Fields of DeprecatedCreateMasterEdition

0: CreateMasterEditionArgs

DeprecatedMintNewEditionFromMasterEditionViaPrintingToken

Given an authority token minted by the Printing mint of a master edition, and a brand new non-metadata-ed mint with one token make a new Metadata + Edition that is a child of the master edition denoted by this authority token. 0. [writable] New Metadata key (pda of [‘metadata’, program id, mint id])

[writable] New Edition V1 (pda of [‘metadata’, program id, mint id, ‘edition’])

[writable] Master Record Edition V1 (pda of [‘metadata’, program id, master metadata mint id, ‘edition’])

[writable] Mint of new token - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[signer] Mint authority of new mint

[writable] Printing Mint of master record edition

[writable] Token account containing Printing mint token to be transferred

[writable] Edition pda to mark creation - will be checked for pre-existence. (pda of [‘metadata’, program id, master mint id, edition_number])

[signer] Burn authority for this token

[signer] payer

[] update authority info for new metadata account

[] Master record metadata account

[] Token program

[] System program

[] Rent info

[optional/writable] Reservation List - If present, and you are on this list, you can get an edition number given by your position on the list.

UpdatePrimarySaleHappenedViaToken

Allows updating the primary sale boolean on Metadata solely through owning an account containing a token from the metadata’s mint and being a signer on this transaction. A sort of limited authority for limited update capability that is required for things like Metaplex to work without needing full authority passing.

[writable] Metadata key (pda of [‘metadata’, program id, mint id])

[signer] Owner on the token account

[] Account containing tokens from the metadata’s mint

DeprecatedSetReservationList(SetReservationListArgs)

Reserve up to 200 editions in sequence for up to 200 addresses in an existing reservation PDA, which can then be used later by redeemers who have printing tokens as a reservation to get a specific edition number as opposed to whatever one is currently listed on the master edition. Used by Auction Manager to guarantee printing order on bid redemption. AM will call whenever the first person redeems a printing bid to reserve the whole block of winners in order and then each winner when they get their token submits their mint and account with the pda that was created by that first bidder - the token metadata can then cross reference these people with the list and see that bidder A gets edition #2, so on and so forth.

NOTE: If you have more than 20 addresses in a reservation list, this may be called multiple times to build up the list, otherwise, it simply wont fit in one transaction. Only provide a total_reservation argument on the first call, which will allocate the edition space, and in follow up calls this will specifically be unnecessary (and indeed will error.)

[writable] Master Edition V1 key (pda of [‘metadata’, program id, mint id, ‘edition’])

[writable] PDA for ReservationList of [‘metadata’, program id, master edition key, ‘reservation’, resource-key]

[signer] The resource you tied the reservation list too

Tuple Fields of DeprecatedSetReservationList

0: SetReservationListArgs

DeprecatedCreateReservationList

Create an empty reservation list for a resource who can come back later as a signer and fill the reservation list with reservations to ensure that people who come to get editions get the number they expect. See SetReservationList for more.

[writable] PDA for ReservationList of [‘metadata’, program id, master edition key, ‘reservation’, resource-key]

[signer] Payer

[signer] Update authority

[] Master Edition V1 key (pda of [‘metadata’, program id, mint id, ‘edition’])

[] A resource you wish to tie the reservation list to. This is so your later visitors who come to redeem can derive your reservation list PDA with something they can easily get at. You choose what this should be.

[] Metadata key (pda of [‘metadata’, program id, mint id])

[] System program

[] Rent info

SignMetadata

Sign a piece of metadata that has you as an unverified creator so that it is now verified. 0. [writable] Metadata (pda of [‘metadata’, program id, mint id])

[signer] Creator

DeprecatedMintPrintingTokensViaToken(MintPrintingTokensViaTokenArgs)

Using a one time authorization token from a master edition v1, print any number of printing tokens from the printing_mint one time, burning the one time authorization token.

[writable] Destination account

[writable] Token account containing one time authorization token

[writable] One time authorization mint

[writable] Printing mint

[signer] Burn authority

[] Metadata key (pda of [‘metadata’, program id, mint id])

[] Master Edition V1 key (pda of [‘metadata’, program id, mint id, ‘edition’])

[] Token program

[] Rent

Tuple Fields of DeprecatedMintPrintingTokensViaToken

0: MintPrintingTokensViaTokenArgs

DeprecatedMintPrintingTokens(MintPrintingTokensViaTokenArgs)

Using your update authority, mint printing tokens for your master edition.

[writable] Destination account

[writable] Printing mint

[signer] Update authority

[] Metadata key (pda of [‘metadata’, program id, mint id])

[] Master Edition V1 key (pda of [‘metadata’, program id, mint id, ‘edition’])

[] Token program

[] Rent

Tuple Fields of DeprecatedMintPrintingTokens

0: MintPrintingTokensViaTokenArgs

CreateMasterEdition(CreateMasterEditionArgs)

Register a Metadata as a Master Edition V2, which means Edition V2s can be minted. Henceforth, no further tokens will be mintable from this primary mint. Will throw an error if more than one token exists, and will throw an error if less than one token exists in this primary mint. 0. [writable] Unallocated edition V2 account with address as pda of [‘metadata’, program id, mint, ‘edition’]

[writable] Metadata mint

[signer] Update authority

[signer] Mint authority on the metadata’s mint - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[signer] payer

[] Metadata account

[] Token program

[] System program

[] Rent info

Tuple Fields of CreateMasterEdition

0: CreateMasterEditionArgs

MintNewEditionFromMasterEditionViaToken(MintNewEditionFromMasterEditionViaTokenArgs)

Given a token account containing the master edition token to prove authority, and a brand new non-metadata-ed mint with one token make a new Metadata + Edition that is a child of the master edition denoted by this authority token. 0. [writable] New Metadata key (pda of [‘metadata’, program id, mint id])

[writable] New Edition (pda of [‘metadata’, program id, mint id, ‘edition’])

[writable] Master Record Edition V2 (pda of [‘metadata’, program id, master metadata mint id, ‘edition’])

[writable] Mint of new token - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[writable] Edition pda to mark creation - will be checked for pre-existence. (pda of [‘metadata’, program id, master metadata mint id, ‘edition’, edition_number]) where edition_number is NOT the edition number you pass in args but actually edition_number = floor(edition/EDITION_MARKER_BIT_SIZE).

[signer] Mint authority of new mint

[signer] payer

[signer] owner of token account containing master token (#8)

[] token account containing token from master metadata mint

[] Update authority info for new metadata

[] Master record metadata account

[] Token program

[] System program

[] Rent info

Tuple Fields of MintNewEditionFromMasterEditionViaToken

0: MintNewEditionFromMasterEditionViaTokenArgs

ConvertMasterEditionV1ToV2

Converts the Master Edition V1 to a Master Edition V2, draining lamports from the two printing mints to the owner of the token account holding the master edition token. Permissionless. Can only be called if there are currenly no printing tokens or one time authorization tokens in circulation.

[writable] Master Record Edition V1 (pda of [‘metadata’, program id, master metadata mint id, ‘edition’])

[writable] One time authorization mint

[writable] Printing mint

MintNewEditionFromMasterEditionViaVaultProxy(MintNewEditionFromMasterEditionViaTokenArgs)

Proxy Call to Mint Edition using a Store Token Account as a Vault Authority.

[writable] New Metadata key (pda of [‘metadata’, program id, mint id])

[writable] New Edition (pda of [‘metadata’, program id, mint id, ‘edition’])

[writable] Master Record Edition V2 (pda of [‘metadata’, program id, master metadata mint id, ‘edition’]

[writable] Mint of new token - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[writable] Edition pda to mark creation - will be checked for pre-existence. (pda of [‘metadata’, program id, master metadata mint id, ‘edition’, edition_number]) where edition_number is NOT the edition number you pass in args but actually edition_number = floor(edition/EDITION_MARKER_BIT_SIZE).

[signer] Mint authority of new mint

[signer] payer

[signer] Vault authority

[] Safety deposit token store account

[] Safety deposit box

[] Vault

[] Update authority info for new metadata

[] Master record metadata account

[] Token program

[] Token vault program

[] System program

[] Rent info

Tuple Fields of MintNewEditionFromMasterEditionViaVaultProxy

0: MintNewEditionFromMasterEditionViaTokenArgs

PuffMetadata

Puff a Metadata - make all of it’s variable length fields (name/uri/symbol) a fixed length using a null character so that it can be found using offset searches by the RPC to make client lookups cheaper. 0. [writable] Metadata account

UpdateMetadataAccountV2(UpdateMetadataAccountArgsV2)

Update a Metadata with is_mutable as a parameter 0. [writable] Metadata account

[signer] Update authority key

Tuple Fields of UpdateMetadataAccountV2

0: UpdateMetadataAccountArgsV2

CreateMetadataAccountV2(CreateMetadataAccountArgsV2)

Create Metadata object. 0. [writable] Metadata key (pda of [‘metadata’, program id, mint id])

[] Mint of token asset

[signer] Mint authority

[signer] payer

[] update authority info

[] System program

[] Rent info

Tuple Fields of CreateMetadataAccountV2

0: CreateMetadataAccountArgsV2

CreateMasterEditionV3(CreateMasterEditionArgs)

Register a Metadata as a Master Edition V2, which means Edition V2s can be minted. Henceforth, no further tokens will be mintable from this primary mint. Will throw an error if more than one token exists, and will throw an error if less than one token exists in this primary mint. 0. [writable] Unallocated edition V2 account with address as pda of [‘metadata’, program id, mint, ‘edition’]

[writable] Metadata mint

[signer] Update authority

[signer] Mint authority on the metadata’s mint - THIS WILL TRANSFER AUTHORITY AWAY FROM THIS KEY

[signer] payer

[writable] Metadata account

[] Token program

[] System program

[] Rent info

Tuple Fields of CreateMasterEditionV3

0: CreateMasterEditionArgs

VerifyCollection

See verify_collection for Doc

Utilize(UtilizeArgs)

See utilize for Doc

Tuple Fields of Utilize

0: UtilizeArgs

ApproveUseAuthority(ApproveUseAuthorityArgs)

See approve_use_authority for Doc

Tuple Fields of ApproveUseAuthority

0: ApproveUseAuthorityArgs

RevokeUseAuthority

See revoke_use_authority for Doc

UnverifyCollection