Module mpl_token_metadata::processor

Functions

process_convert_master_edition_v1_to_v2

process_create_master_edition

Create master edition

process_create_metadata_accounts_v2

process_instruction

process_mint_new_edition_from_master_edition_via_token

process_mint_new_edition_from_master_edition_via_vault_proxy

process_puff_metadata_account

Puff out the variable length fields to a fixed length on a metadata account in a permissionless way.

process_sign_metadata

process_update_metadata_accounts_v2

process_update_primary_sale_happened_via_token

verify_collection