Module mpl_token_metadata::state

Structs

Collection

Creator

Data

DataV2

Edition

All Editions should never have a supply greater than 1. To enforce this, a transfer mint authority instruction will happen when a normal token is turned into an Edition, and in order for a Metadata update authority to do this transaction they will also need to sign the transaction as the Mint authority.

EditionMarker

MasterEditionV1

MasterEditionV2

Metadata

Reservation

ReservationListV1

ReservationListV2

ReservationV1

Uses

Enums

Key

TokenStandard

UseMethod

Constants

EDITION

Used in seeds to make Edition model pda address

EDITION_MARKER_BIT_SIZE

MAX_CREATOR_LEN

MAX_CREATOR_LIMIT

MAX_DATA_SIZE

MAX_EDITION_LEN

MAX_EDITION_MARKER_SIZE

MAX_MASTER_EDITION_LEN

MAX_METADATA_LEN

MAX_NAME_LENGTH

MAX_RESERVATIONS

MAX_RESERVATION_LIST_SIZE

MAX_RESERVATION_LIST_V1_SIZE

MAX_SYMBOL_LENGTH

MAX_URI_LENGTH

PREFIX

prefix used for PDAs to avoid certain collision attacks (https://en.wikipedia.org/wiki/Collision_attack#Chosen-prefix_collision_attack)

RESERVATION

USER

Traits

MasterEdition

ReservationList

Functions

get_master_edition

get_reservation_list