Struct mpl_token_metadata::state::ReservationListV2

#[repr(C)]
pub struct ReservationListV2 {
    pub key: Key,
    pub master_edition: Pubkey,
    pub supply_snapshot: Option<u64>,
    pub reservations: Vec<Reservation>,
    pub total_reservation_spots: u64,
    pub current_reservation_spots: u64,
}

Fields
key: Keymaster_edition: Pubkey

Present for reverse lookups

supply_snapshot: Option<u64>

What supply counter was on master_edition when this reservation was created.

reservations: Vec<Reservation>total_reservation_spots: u64

How many reservations there are going to be, given on first set_reservation call

current_reservation_spots: u64

Cached count of reservation spots in the reservation vec to save on CPU.

Implementations

impl ReservationListV2

pub fn from_account_info(
a: &AccountInfo<'_>
) -> Result<ReservationListV2, ProgramError>