Module mpl_token_metadata::utils

Structs

CreateMetadataAccountsLogicArgs

MintNewEditionFromMasterEditionViaTokenLogicArgs

TokenBurnParams

TokenBurnParams

TokenMintToParams

TokenMintToParams

Functions

assert_currently_holding

assert_data_valid

assert_derivation

assert_edition_valid

assert_initialized

assert initialized account

assert_mint_authority_matches_mint

assert_owned_by

assert_rent_exempt

assert_signer

assert_supply_invariance

assert_token_program_matches_package

assert_update_authority_is_correct

calculate_edition_number

calculate_supply_change

create_or_allocate_account_raw

Create account almost from scratch, lifted from https://github.com/solana-labs/solana-program-library/tree/master/associated-token-account/program/src/processor.rs#L51-L98

extract_edition_number_from_deprecated_reservation_list

get_mint_authority

get_mint_decimals

cheap method to just get supply off a mint without unpacking whole object

get_mint_freeze_authority

get_mint_supply

cheap method to just get supply off a mint without unpacking whole object

get_owner_from_token_account

Cheap method to just grab owner Pubkey from token account, instead of deserializing entire thing

get_supply_off_master_edition

mint_limited_edition

process_create_metadata_accounts_logic

Create a new account instruction

process_mint_new_edition_from_master_edition_via_token_logic

puff_out_data_fields

Strings need to be appended with \0s in order to have a deterministic length. This supports the memcmp filter on get program account calls. NOTE: it is assumed that the metadata fields are never larger than the respective MAX_LENGTH

puffed_out_string

Pads the string to the desired size with 0u8s. NOTE: it is assumed that the string’s size is never larger than the given size.

spl_token_burn

spl_token_mint_to

transfer_mint_authority

try_from_slice_checked

zero_account

Pads the string to the desired size with 0u8s. NOTE: it is assumed that the string’s size is never larger than the given size.