mpl_token_metadata::errors
Enum MplTokenMetadataError

pub enum MplTokenMetadataError {

    InstructionUnpackError,
    InstructionPackError,
    NotRentExempt,
    AlreadyInitialized,
    Uninitialized,
    InvalidMetadataKey,
    InvalidEditionKey,
    UpdateAuthorityIncorrect,
    UpdateAuthorityIsNotSigner,
    NotMintAuthority,
    InvalidMintAuthority,
    NameTooLong,
    SymbolTooLong,
    UriTooLong,
    UpdateAuthorityMustBeEqualToMetadataAuthorityAndSigner,
    MintMismatch,
    EditionsMustHaveExactlyOneToken,
    MaxEditionsMintedAlready,
    TokenMintToFailed,
    MasterRecordMismatch,
    DestinationMintMismatch,
    EditionAlreadyMinted,
    PrintingMintDecimalsShouldBeZero,
    OneTimePrintingAuthorizationMintDecimalsShouldBeZero,
    EditionMintDecimalsShouldBeZero,
    TokenBurnFailed,
    TokenAccountOneTimeAuthMintMismatch,
    DerivedKeyInvalid,
    PrintingMintMismatch,
    OneTimePrintingAuthMintMismatch,
    TokenAccountMintMismatch,
    TokenAccountMintMismatchV2,
    NotEnoughTokens,
    PrintingMintAuthorizationAccountMismatch,
    AuthorizationTokenAccountOwnerMismatch,
    Disabled,
    CreatorsTooLong,
    CreatorsMustBeAtleastOne,
    MustBeOneOfCreators,
    NoCreatorsPresentOnMetadata,
    CreatorNotFound,
    InvalidBasisPoints,
    PrimarySaleCanOnlyBeFlippedToTrue,
    OwnerMismatch,
    NoBalanceInAccountForAuthorization,
    ShareTotalMustBe100,
    ReservationExists,
    ReservationDoesNotExist,
    ReservationNotSet,
    ReservationAlreadyMade,
    BeyondMaxAddressSize,
    NumericalOverflowError,
    ReservationBreachesMaximumSupply,
    AddressNotInReservation,
    CannotVerifyAnotherCreator,
    CannotUnverifyAnotherCreator,
    SpotMismatch,
    IncorrectOwner,
    PrintingWouldBreachMaximumSupply,
    DataIsImmutable,
    DuplicateCreatorAddress,
    ReservationSpotsRemainingShouldMatchTotalSpotsAtStart,
    InvalidTokenProgram,
    DataTypeMismatch,
    BeyondAlottedAddressSize,
    ReservationNotComplete,
    TriedToReplaceAnExistingReservation,
    InvalidOperation,
    InvalidOwner,
    PrintingMintSupplyMustBeZeroForConversion,
    OneTimeAuthMintSupplyMustBeZeroForConversion,
    InvalidEditionIndex,
    ReservationArrayShouldBeSizeOne,
    IsMutableCanOnlyBeFlippedToFalse,
    CollectionCannotBeVerifiedInThisInstruction,
    Removed,
    MustBeBurned,
    InvalidUseMethod,
    CannotChangeUseMethodAfterFirstUse,
    CannotChangeUsesAfterFirstUse,
    CollectionNotFound,
    InvalidCollectionUpdateAuthority,
    CollectionMustBeAUniqueMasterEdition,
    UseAuthorityRecordAlreadyExists,
    UseAuthorityRecordAlreadyRevoked,
    Unusable,
    NotEnoughUses,
    CollectionAuthorityRecordAlreadyExists,
    CollectionAuthorityDoesNotExist,
    InvalidUseAuthorityRecord,
    InvalidCollectionAuthorityRecord,
    InvalidFreezeAuthority,
    InvalidDelegate,
    CannotAdjustVerifiedCreator,
    CannotRemoveVerifiedCreator,
    CannotWipeVerifiedCreators,
    NotAllowedToChangeSellerFeeBasisPoints,
    EditionOverrideCannotBeZero,
    InvalidUser,
    RevokeCollectionAuthoritySignerIncorrect,
    TokenCloseFailed,
    UnsizedCollection,
    SizedCollection,
    MissingCollectionMetadata,
    NotAMemberOfCollection,
    NotVerifiedMemberOfCollection,
    NotACollectionParent,
    CouldNotDetermineTokenStandard,
    MissingEditionAccount,
    NotAMasterEdition,
    MasterEditionHasPrints,
    BorshDeserializationError,
    CannotUpdateVerifiedCollection,
    CollectionMasterEditionAccountInvalid,
    AlreadyVerified,
    AlreadyUnverified,
    NotAPrintEdition,
    InvalidMasterEdition,
    InvalidPrintEdition,
    InvalidEditionMarker,
    ReservationListDeprecated,
    PrintEditionDoesNotMatchMasterEdition,
    EditionNumberGreaterThanMaxSupply,
    MustUnverify,
    InvalidEscrowBumpSeed,
    MustBeEscrowAuthority,
    InvalidSystemProgram,
    MustBeNonFungible,
    InsufficientTokens,
    BorshSerializationError,
    NoFreezeAuthoritySet,
    InvalidCollectionSizeChange,
    InvalidBubblegumSigner,
    EscrowParentHasDelegate,
    MintIsNotSigner,
    InvalidTokenStandard,
    InvalidMintForTokenStandard,
    InvalidAuthorizationRules,
    MissingAuthorizationRules,
    MissingProgrammableConfig,
    InvalidProgrammableConfig,
    DelegateAlreadyExists,
    DelegateNotFound,
    MissingAccountInBuilder,
    MissingArgumentInBuilder,
    FeatureNotSupported,
    InvalidSystemWallet,
    OnlySaleDelegateCanTransfer,
    MissingTokenAccount,
    MissingSplTokenProgram,
    MissingAuthorizationRulesProgram,
    InvalidDelegateRoleForTransfer,
    InvalidTransferAuthority,
    InstructionNotSupported,
    KeyMismatch,
    LockedToken,
    UnlockedToken,
    MissingDelegateRole,
    InvalidAuthorityType,
    MissingTokenRecord,
    MintSupplyMustBeZero,
    DataIsEmptyOrZeroed,
    MissingTokenOwnerAccount,
    InvalidMasterEditionAccountLength,
    IncorrectTokenState,
    InvalidDelegateRole,
    MissingPrintSupply,
    MissingMasterEditionAccount,
    AmountMustBeGreaterThanZero,
    InvalidDelegateArgs,
    MissingLockedTransferAddress,
    InvalidLockedTransferAddress,
    DataIncrementLimitExceeded,
    CannotUpdateAssetWithDelegate,
    InvalidAmount,
    MissingMasterEditionMintAccount,
    MissingMasterEditionTokenAccount,
    MissingEditionMarkerAccount,
    CannotBurnWithDelegate,
    MissingEdition,
    InvalidAssociatedTokenAccountProgram,
    InvalidInstructionsSysvar,
    InvalidParentAccounts,
    InvalidUpdateArgs,
    InsufficientTokenBalance,
    MissingCollectionMint,
    MissingCollectionMasterEdition,
    InvalidTokenRecord,
    InvalidCloseAuthority,
    InvalidInstruction,
    MissingDelegateRecord,
    InvalidFeeAccount,
    InvalidMetadataFlags,
    CannotChangeUpdateAuthorityWithDelegate,
    InvalidMintExtensionType,
    InvalidMintCloseAuthority,
    InvalidMetadataPointer,
    InvalidTokenExtensionType,
    MissingImmutableOwnerExtension,
    ExpectedUninitializedAccount,
    InvalidEditionAccountLength,
    AccountAlreadyResized,
    ConditionsForClosingNotMet,

}

Variants

InstructionUnpackError

0 (0x0) -

InstructionPackError

1 (0x1) -

NotRentExempt

2 (0x2) - Lamport balance below rent-exempt threshold

AlreadyInitialized

3 (0x3) - Already initialized

Uninitialized

4 (0x4) - Uninitialized

InvalidMetadataKey

5 (0x5) - Metadata’s key must match seed of [‘metadata’, program id, mint] provided

InvalidEditionKey

6 (0x6) - Edition’s key must match seed of [‘metadata’, program id, name, ‘edition’] provided

UpdateAuthorityIncorrect

7 (0x7) - Update Authority given does not match

UpdateAuthorityIsNotSigner

8 (0x8) - Update Authority needs to be signer to update metadata

NotMintAuthority

9 (0x9) - You must be the mint authority and signer on this transaction

InvalidMintAuthority

10 (0xA) - Mint authority provided does not match the authority on the mint

NameTooLong

11 (0xB) - Name too long

SymbolTooLong

12 (0xC) - Symbol too long

UriTooLong

13 (0xD) - URI too long

UpdateAuthorityMustBeEqualToMetadataAuthorityAndSigner

14 (0xE) -

MintMismatch

15 (0xF) - Mint given does not match mint on Metadata

EditionsMustHaveExactlyOneToken

16 (0x10) - Editions must have exactly one token

MaxEditionsMintedAlready

17 (0x11) -

TokenMintToFailed

18 (0x12) -

MasterRecordMismatch

19 (0x13) -

DestinationMintMismatch

20 (0x14) -

EditionAlreadyMinted

21 (0x15) -

PrintingMintDecimalsShouldBeZero

22 (0x16) -

OneTimePrintingAuthorizationMintDecimalsShouldBeZero

23 (0x17) -

EditionMintDecimalsShouldBeZero

24 (0x18) - EditionMintDecimalsShouldBeZero

TokenBurnFailed

25 (0x19) -

TokenAccountOneTimeAuthMintMismatch

26 (0x1A) -

DerivedKeyInvalid

27 (0x1B) - Derived key invalid

PrintingMintMismatch

28 (0x1C) - The Printing mint does not match that on the master edition!

OneTimePrintingAuthMintMismatch

29 (0x1D) - The One Time Printing Auth mint does not match that on the master edition!

TokenAccountMintMismatch

30 (0x1E) - The mint of the token account does not match the Printing mint!

TokenAccountMintMismatchV2

31 (0x1F) - The mint of the token account does not match the master metadata mint!

NotEnoughTokens

32 (0x20) - Not enough tokens to mint a limited edition

PrintingMintAuthorizationAccountMismatch

33 (0x21) -

AuthorizationTokenAccountOwnerMismatch

34 (0x22) -

Disabled

35 (0x23) -

CreatorsTooLong

36 (0x24) - Creators list too long

CreatorsMustBeAtleastOne

37 (0x25) - Creators must be at least one if set

MustBeOneOfCreators

38 (0x26) -

NoCreatorsPresentOnMetadata

39 (0x27) - This metadata does not have creators

CreatorNotFound

40 (0x28) - This creator address was not found

InvalidBasisPoints

41 (0x29) - Basis points cannot be more than 10000

PrimarySaleCanOnlyBeFlippedToTrue

42 (0x2A) - Primary sale can only be flipped to true and is immutable

OwnerMismatch

43 (0x2B) - Owner does not match that on the account given

NoBalanceInAccountForAuthorization

44 (0x2C) - This account has no tokens to be used for authorization

ShareTotalMustBe100

45 (0x2D) - Share total must equal 100 for creator array

ReservationExists

46 (0x2E) -

ReservationDoesNotExist

47 (0x2F) -

ReservationNotSet

48 (0x30) -

ReservationAlreadyMade

49 (0x31) -

BeyondMaxAddressSize

50 (0x32) -

NumericalOverflowError

51 (0x33) - NumericalOverflowError

ReservationBreachesMaximumSupply

52 (0x34) -

AddressNotInReservation

53 (0x35) -

CannotVerifyAnotherCreator

54 (0x36) - You cannot unilaterally verify another creator, they must sign

CannotUnverifyAnotherCreator

55 (0x37) - You cannot unilaterally unverify another creator

SpotMismatch

56 (0x38) -

IncorrectOwner

57 (0x39) - Incorrect account owner

PrintingWouldBreachMaximumSupply

58 (0x3A) -

DataIsImmutable

59 (0x3B) - Data is immutable

DuplicateCreatorAddress

60 (0x3C) - No duplicate creator addresses

ReservationSpotsRemainingShouldMatchTotalSpotsAtStart

61 (0x3D) -

InvalidTokenProgram

62 (0x3E) - Invalid token program

DataTypeMismatch

63 (0x3F) - Data type mismatch

BeyondAlottedAddressSize

64 (0x40) -

ReservationNotComplete

65 (0x41) -

TriedToReplaceAnExistingReservation

66 (0x42) -

InvalidOperation

67 (0x43) - Invalid operation

InvalidOwner

68 (0x44) - Invalid Owner

PrintingMintSupplyMustBeZeroForConversion

69 (0x45) - Printing mint supply must be zero for conversion

OneTimeAuthMintSupplyMustBeZeroForConversion

70 (0x46) - One Time Auth mint supply must be zero for conversion

InvalidEditionIndex

71 (0x47) - You tried to insert one edition too many into an edition mark pda

ReservationArrayShouldBeSizeOne

72 (0x48) -

IsMutableCanOnlyBeFlippedToFalse

73 (0x49) - Is Mutable can only be flipped to false

CollectionCannotBeVerifiedInThisInstruction

74 (0x4A) - Collection cannot be verified in this instruction

Removed

75 (0x4B) - This instruction was deprecated in a previous release and is now removed

MustBeBurned

76 (0x4C) -

InvalidUseMethod

77 (0x4D) - This use method is invalid

CannotChangeUseMethodAfterFirstUse

78 (0x4E) - Cannot Change Use Method after the first use

CannotChangeUsesAfterFirstUse

79 (0x4F) - Cannot Change Remaining or Available uses after the first use

CollectionNotFound

80 (0x50) - Collection Not Found on Metadata

InvalidCollectionUpdateAuthority

81 (0x51) - Collection Update Authority is invalid

CollectionMustBeAUniqueMasterEdition

82 (0x52) - Collection Must Be a Unique Master Edition v2

UseAuthorityRecordAlreadyExists

83 (0x53) - The Use Authority Record Already Exists, to modify it Revoke, then Approve

UseAuthorityRecordAlreadyRevoked

84 (0x54) - The Use Authority Record is empty or already revoked

Unusable

85 (0x55) - This token has no uses

NotEnoughUses

86 (0x56) - There are not enough Uses left on this token.

CollectionAuthorityRecordAlreadyExists

87 (0x57) - This Collection Authority Record Already Exists.

CollectionAuthorityDoesNotExist

88 (0x58) - This Collection Authority Record Does Not Exist.

InvalidUseAuthorityRecord

89 (0x59) - This Use Authority Record is invalid.

InvalidCollectionAuthorityRecord

90 (0x5A) -

InvalidFreezeAuthority

91 (0x5B) - Metadata does not match the freeze authority on the mint

InvalidDelegate

92 (0x5C) - All tokens in this account have not been delegated to this user.

CannotAdjustVerifiedCreator

93 (0x5D) -

CannotRemoveVerifiedCreator

94 (0x5E) - Verified creators cannot be removed.

CannotWipeVerifiedCreators

95 (0x5F) -

NotAllowedToChangeSellerFeeBasisPoints

96 (0x60) -

EditionOverrideCannotBeZero

97 (0x61) - Edition override cannot be zero

InvalidUser

98 (0x62) - Invalid User

RevokeCollectionAuthoritySignerIncorrect

99 (0x63) - Revoke Collection Authority signer is incorrect

TokenCloseFailed

100 (0x64) -

UnsizedCollection

101 (0x65) - Can’t use this function on unsized collection

SizedCollection

102 (0x66) - Can’t use this function on a sized collection

MissingCollectionMetadata

103 (0x67) - Missing collection metadata account

NotAMemberOfCollection

104 (0x68) - This NFT is not a member of the specified collection.

NotVerifiedMemberOfCollection

105 (0x69) - This NFT is not a verified member of the specified collection.

NotACollectionParent

106 (0x6A) - This NFT is not a collection parent NFT.

CouldNotDetermineTokenStandard

107 (0x6B) - Could not determine a TokenStandard type.

MissingEditionAccount

108 (0x6C) - This mint account has an edition but none was provided.

NotAMasterEdition

109 (0x6D) - This edition is not a Master Edition

MasterEditionHasPrints

110 (0x6E) - This Master Edition has existing prints

BorshDeserializationError

111 (0x6F) -

CannotUpdateVerifiedCollection

112 (0x70) - Cannot update a verified collection in this command

CollectionMasterEditionAccountInvalid

113 (0x71) - Edition account doesnt match collection

AlreadyVerified

114 (0x72) - Item is already verified.

AlreadyUnverified

115 (0x73) -

NotAPrintEdition

116 (0x74) - This edition is not a Print Edition

InvalidMasterEdition

117 (0x75) - Invalid Master Edition

InvalidPrintEdition

118 (0x76) - Invalid Print Edition

InvalidEditionMarker

119 (0x77) - Invalid Edition Marker

ReservationListDeprecated

120 (0x78) - Reservation List is Deprecated

PrintEditionDoesNotMatchMasterEdition

121 (0x79) - Print Edition does not match Master Edition

EditionNumberGreaterThanMaxSupply

122 (0x7A) - Edition Number greater than max supply

MustUnverify

123 (0x7B) - Must unverify before migrating collections.

InvalidEscrowBumpSeed

124 (0x7C) - Invalid Escrow Account Bump Seed

MustBeEscrowAuthority

125 (0x7D) - Must Escrow Authority

InvalidSystemProgram

126 (0x7E) - Invalid System Program

MustBeNonFungible

127 (0x7F) - Must be a Non Fungible Token

InsufficientTokens

128 (0x80) - Insufficient tokens for transfer

BorshSerializationError

129 (0x81) - Borsh Serialization Error

NoFreezeAuthoritySet

130 (0x82) - Cannot create NFT with no Freeze Authority.

InvalidCollectionSizeChange

131 (0x83) - Invalid collection size change

InvalidBubblegumSigner

132 (0x84) - Invalid bubblegum signer

EscrowParentHasDelegate

133 (0x85) - Escrow parent cannot have a delegate

MintIsNotSigner

134 (0x86) - Mint needs to be signer to initialize the account

InvalidTokenStandard

135 (0x87) - Invalid token standard

InvalidMintForTokenStandard

136 (0x88) - Invalid mint account for specified token standard

InvalidAuthorizationRules

137 (0x89) - Invalid authorization rules account

MissingAuthorizationRules

138 (0x8A) - Missing authorization rules account

MissingProgrammableConfig

139 (0x8B) - Missing programmable configuration

InvalidProgrammableConfig

140 (0x8C) - Invalid programmable configuration

DelegateAlreadyExists

141 (0x8D) - Delegate already exists

DelegateNotFound

142 (0x8E) - Delegate not found

MissingAccountInBuilder

143 (0x8F) - Required account not set in instruction builder

MissingArgumentInBuilder

144 (0x90) - Required argument not set in instruction builder

FeatureNotSupported

145 (0x91) - Feature not supported currently

InvalidSystemWallet

146 (0x92) - Invalid system wallet

OnlySaleDelegateCanTransfer

147 (0x93) - Only the sale delegate can transfer while its set

MissingTokenAccount

148 (0x94) - Missing token account

MissingSplTokenProgram

149 (0x95) - Missing SPL token program

MissingAuthorizationRulesProgram

150 (0x96) - Missing authorization rules program

InvalidDelegateRoleForTransfer

151 (0x97) - Invalid delegate role for transfer

InvalidTransferAuthority

152 (0x98) - Invalid transfer authority

InstructionNotSupported

153 (0x99) - Instruction not supported for ProgrammableNonFungible assets

KeyMismatch

154 (0x9A) - Public key does not match expected value

LockedToken

155 (0x9B) - Token is locked

UnlockedToken

156 (0x9C) - Token is unlocked

MissingDelegateRole

157 (0x9D) - Missing delegate role

InvalidAuthorityType

158 (0x9E) - Invalid authority type

MissingTokenRecord

159 (0x9F) - Missing token record account

MintSupplyMustBeZero

160 (0xA0) - Mint supply must be zero for programmable assets

DataIsEmptyOrZeroed

161 (0xA1) - Data is empty or zeroed

MissingTokenOwnerAccount

162 (0xA2) - Missing token owner

InvalidMasterEditionAccountLength

163 (0xA3) - Master edition account has an invalid length

IncorrectTokenState

164 (0xA4) - Incorrect token state

InvalidDelegateRole

165 (0xA5) - Invalid delegate role

MissingPrintSupply

166 (0xA6) - Print supply is required for non-fungibles

MissingMasterEditionAccount

167 (0xA7) - Missing master edition account

AmountMustBeGreaterThanZero

168 (0xA8) - Amount must be greater than zero

InvalidDelegateArgs

169 (0xA9) - Invalid delegate args

MissingLockedTransferAddress

170 (0xAA) - Missing address for locked transfer

InvalidLockedTransferAddress

171 (0xAB) - Invalid destination address for locked transfer

DataIncrementLimitExceeded

172 (0xAC) - Exceeded account realloc increase limit

CannotUpdateAssetWithDelegate

173 (0xAD) - Cannot update the rule set of a programmable asset that has a delegate

InvalidAmount

174 (0xAE) - Invalid token amount for this operation or token standard

MissingMasterEditionMintAccount

175 (0xAF) - Missing master edition mint account

MissingMasterEditionTokenAccount

176 (0xB0) - Missing master edition token account

MissingEditionMarkerAccount

177 (0xB1) - Missing edition marker account

CannotBurnWithDelegate

178 (0xB2) - Cannot burn while persistent delegate is set

MissingEdition

179 (0xB3) - Missing edition account

InvalidAssociatedTokenAccountProgram

180 (0xB4) - Invalid Associated Token Account Program

InvalidInstructionsSysvar

181 (0xB5) - Invalid InstructionsSysvar

InvalidParentAccounts

182 (0xB6) - Invalid or Unneeded parent accounts

InvalidUpdateArgs

183 (0xB7) - Authority cannot apply all update args

InsufficientTokenBalance

184 (0xB8) - Token account does not have enough tokens

MissingCollectionMint

185 (0xB9) - Missing collection account

MissingCollectionMasterEdition

186 (0xBA) - Missing collection master edition account

InvalidTokenRecord

187 (0xBB) - Invalid token record account

InvalidCloseAuthority

188 (0xBC) - The close authority needs to be revoked by the Utility Delegate

InvalidInstruction

189 (0xBD) - Invalid or removed instruction

MissingDelegateRecord

190 (0xBE) - Missing delegate record

InvalidFeeAccount

191 (0xBF) -

InvalidMetadataFlags

192 (0xC0) -

CannotChangeUpdateAuthorityWithDelegate

193 (0xC1) - Cannot change the update authority with a delegate

InvalidMintExtensionType

194 (0xC2) - Invalid mint extension type

InvalidMintCloseAuthority

195 (0xC3) - Invalid mint close authority

InvalidMetadataPointer

196 (0xC4) - Invalid metadata pointer

InvalidTokenExtensionType

197 (0xC5) - Invalid token extension type

MissingImmutableOwnerExtension

198 (0xC6) - Missing immutable owner extension

ExpectedUninitializedAccount

199 (0xC7) - Expected account to be uninitialized

InvalidEditionAccountLength

200 (0xC8) - Edition account has an invalid length

AccountAlreadyResized

201 (0xC9) - Account has already been resized

ConditionsForClosingNotMet

202 (0xCA) - Conditions for closing not met